intelligence/requirements.txt
transformers==4.39.3
accelerate==0.28.0
bitsandbytes==0.42.0

peft==0.10.0

langchain==0.1.16
langgraph==0.0.28
langchain-core==0.1.47
langchain-community==0.0.34

sentencepiece==0.1.99
protobuf==4.25.3
tiktoken==0.5.2

black==24.3.0
isort==5.13.2
redbaron==0.9.2
radon==6.0.1
pylint==3.1.0
mypy==1.9.0
bandit==1.7.6
pytest==8.1.1

jpype1

python-dotenv==1.0.1
rich==13.7.1

redis==5.0.3
langchainhub==0.1.14

jsonschema==4.21.1

packaging==23.2
scipy==1.12.0